backend/pyproject.toml
[project]
name = "sataplan"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.14.1",
    "bcrypt>=4.0.1,<4.1.0",
    "fastapi[standard]>=0.115.6",
    "passlib>=1.7.4,<1.8.0",
    "pathlib>=1.0.1",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.0.1",
    "python-jose>=3.3.0",
    "qrcode[pil]>=8.0",
    "requests>=2.32.3",
    "sqlalchemy-utils>=0.41.2",
    "sqlalchemy>=2.0.36",
    "uvicorn>=0.34.0",
]
